# pillars/contacts/openapi/contacts.openapi.json
{
  "components": {
    "schemas": {
      "Address": {
        "description": "The wire shape served by `GET`/`POST /entities/:id/addresses`.",
        "properties": {
          "entityId": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "lastEditedTime": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "entityId",
          "value",
          "lastEditedTime"
        ],
        "type": "object"
      },
      "AddressListResponse": {
        "description": "`GET /entities/{id}/addresses` response body.",
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Address"
            },
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
      "AddressMutation": {
        "description": "`POST /entities/{id}/addresses` response body.",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Address"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "data",
          "message"
        ],
        "type": "object"
      },
      "CreateAddressBody": {
        "description": "Body accepted by `POST /entities/:id/addresses`.",
        "properties": {
          "value": {
            "type": "string"
          }
        },
        "required": [
          "value"
        ],
        "type": "object"
      },
...
      "ErrorBody": {
        "description": "Error envelope returned by every fallible contacts route.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "details": {},
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "requestId",
          "retryable"
        ],
        "type": "object"
      },
      "ErrorCode": {
        "description": "Stable error codes returned by the contacts pillar.",
        "enum": [
          "contacts.request.invalid",
          "contacts.resource.not_found",
          "contacts.entity.name_conflict",
          "contacts.internal"
        ],
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "Contacts pillar — authoritative entities store (first Rust pillar).",
    "license": {
      "identifier": "UNLICENSED",
      "name": "UNLICENSED"
    },
    "title": "POPS Contacts",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/entities/{id}/addresses": {
      "get": {
        "operationId": "entities.addresses.list",
        "parameters": [
          {
            "description": "Entity id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressListResponse"
                }
              }
            },
            "description": "The entity's addresses"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such entity"
          }
        },
        "tags": [
          "crate::entities::addresses_routes"
        ]
      },
      "post": {
        "operationId": "entities.addresses.create",
        "parameters": [
          {
            "description": "Entity id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAddressBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressMutation"
                }
              }
            },
            "description": "Created address"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid body"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such entity"
          }
        },
        "tags": [
          "crate::entities::addresses_routes"
        ]
      }
    },
...
  }
}